+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to **turn_by_turn** will be documented in this file.
 
 ### v1.5.0 - 2026-07-23
 
-Now, when reading PSB files, the horizontal position of the BPM is negated to match the model convention, as the PSB runs anticlockwise.
+Now, when reading PSB files, the horizontal position of the BPM is inverted to match the model convention, as the PSB runs anticlockwise.
 
 
 ### v1.4.1 - 2026-04-13